camkes: General user-level proof about GetMR/SetMR.
These kind of properties are required for future reasoning about glue code
marshalling and unmarshalling. However, almost any lemma we need about libsel4
is desirable for any user-level program on seL4. Nothing is particularly
compositional right now, which isn't too much of a problem for CAmkES where
we're generating proofs, but anyone else who wants to use this is stuck with
copy-and-paste.

JIRA: VER-159