attribute tracing: Mechanism to work out changes in simpsets across revisions.
The idea of this file is to allow users to determine how the simpset,
cong set, intro set, wp sets, etc. have changed from an old version of
the repository to a new version.

The process is as follows:

  1. A user runs "save_attributes" on an old, working version of the
     theory.

  2. This tool will write out a ".foo.attrib_trace" file for each
     theory processed.

  3. The user modifies imports statements as required, possibly
     breaking the proof.

  4. The user can now run "diff_attributes" to determine what
     commands they should run to restore the simpset / congset /etc
     to something closer to the old version.

The tool is not complete, in that it won't always suggest the full set
of "simp add", "simp del", etc commands. Nor does it know that a rule
added to the simpset is causing a problem. It merely lists
a hopefully-sensible set of differences.