regression: Attempt to make "run_tests.py" better at cleaning up after itself.
Currently, when Ctrl+C is pressed, background "poly" processes often
continue to run, even if "pidspace" is being used. Eeew!

We now make a better attempt at killing our children when Ctrl+C is
pressed. In particular, we use the "psutil" python library instead of
the "timeout" binary. The former should be more portable, working on
MacOS X systems as well.

"psutil" and "pidspace" don't interact well; "psutil" doesn't seem to
find children inside the pidspace, which is problematic. Might need to
address this a little more in the future.

This patch dedicated to Matt.