c-parser: Generate vcg rule for "cchaos" constant.
Allows vcg_tag ("apply vcg") to process functions with "cchaos" in them.

This was an attempt to solve the problem that the C parser's modifies
proofs currently fail for functions containing "cchaos".

The problem, however, actually seems to be that the C parser fails to
realise that incorrectly owned _read_ variables (which are updated using
"cchaos" before being read) need to be in the "modifies" set of
a function. Currently, the C parser misses these modifications, and the
proof of the incorrect modifies set (rightly) fails.