autocorres: restore backwards compatibility for heap lifting struct accesses.
Structs may be nested or non-nested, and packed or non-packed.
The old heap_lift worked for all non-nested structs; the rules were recently
changed to work for all packed structs.

This patch simply adds the old rules alongside the new so that
any non-nested or packed struct accesses can be heap lifted.

In the future, we might want to fix the remaining unsupported case:
nested non-packed structs.