autocorres examples: Use the memcpy lemma to prove something about a caller.
Reasoning about heap equivalence involving mixed types (~= writing to memory
through two differently typed pointers) is very awkward. The enclosed proof
leans heavily on the fact that we know the concrete size of the type of the
memory we're copying, and it still requires a fair bit of sledgehammering. This
is fine for an example, but certainly won't translate to proof generation on
arbitrary types. We may need to take a slightly different tack.