autocorres examples: Prove a heap-abstracted call to memcpy.
This commit adds a proof over a call to (the non-heap-abstracted) memcpy from
a heap-abstracted context. As a result, our pre- and post-conditions are
expressed on the abstract heap and need to be transfered to the concrete heap
during the proof. Altogether the proof is not so complex and most of the heavy
lifting remains in the original proof of memcpy.