autocorres examples: Extended proofs for memcpying a struct.
This extra proof was to validate that the existing word 32 proof generalises to
other, arbitrarily complex, types. The proof script is more or less identical,
which strongly suggests that we can now trivially lift the generic memcpy proof
into a proof for copying any mem_type. It would be nice to generalise the
required properties on is_valid_* into some kind of locale construction, but the
way lifted_globals and friends are constructed make this a little tricky.

Note that the memcpy implementation we're currently reasoning about should
actually work correctly even if the memory regions overlap, as long as the
destination pointer is less than or equal to the source. That is, the
precondition of the Hoare triple for memcpy could be weakened. However, this
does not seem sensible to do as the memcpy spec requires that the input
pointers do not overlap and we don't want users proving properties that they
think will also hold on other memcpy implementations.