Handle local static variables properly.
They are now treated as globals with a specially munged name, derived
from the given name and the name of the function where they occur. The
function NameGeneration.mk_localstatic generates the "munged" name.

As with other globals, initialisation is not handled very well (i.e., at
the moment the initialisation is completely ignored).

Close JIRA VER-439