misc/pysymbols: Don't translate entities like "\<^bsub>" into unicode.
The code points of the xsymbols whose ASCII text begins with "\<^" appear to be
a ruse. Their unicode characters do not correspond to their intent because
they are actually format markers. For example, "\<^bsub>" maps to something
described as "south east double arrow," but is intended to mean "start
subscript."

It turns out we do not need to encode these at all because a sufficiently
clever Pygments can interpret their ASCII representations. One such clever
instance is [0]. This commit removes encoding of all symbols whose ASCII
representation begins with "\<^" under the assumption they will be handled by a
later tool in your pipeline.

  [0]: https://bitbucket.org/lsf37/pygments-main/branch/sub-super