Export file containing varname munging info
Depends on configuration variable (as per example in
testfiles/jira473.thy), "munge_info_fname".  If this is not set, then no
export occurs.  File written is with respect to directory Isabelle is
started in, which may not be ideal.

Format in the file is a series of lines:

  function_name::var_name -> munged_name

If the variable is global, then function_name is blank.

JIRA VER-473