ccorres_rewrite: implement linear-time ccorres simplification
The previous implementation took quadratic time in the worst case,
because it repeatedly traversed the term from the top until no more
simplifications are possible.

For linear time, we apply a transitivity rule before each split rule or
simplification. This allows us to completely finish rewriting on
subterms before we reassemble the larger term which comprises them.

However, we must be careful not to apply split rules to terms which are
already simplified. To avoid this, we first split as much as possible,
and then never split again. In the final phase, we simplify where
possible, and otherwise use reflexivity to reassemble.

We also simplify the required form of C_simp rules, for more natural
statements of the intended simplifications.