x64: get StateRelation, Invariants_H building
This will now allow sorrying of proofs in refine to aid parallelism as
we proceed.

Note: the invariants are in no way complete and the aim should be to
update them as required - allowing people to start working on refine
proofs without the high initial cost of fixing/writing all of the
invariants.