lib/corres_method: add more conservative correswp
Correswp is wp but with more conservative treatment for
schematics. Rules in wp_comb that do precondition weakening
are avoided when the precondition is schematic, and there
is a final check which fails if any schematic preconditions
are found.

Realistically this should be the default behaviour for wp, but
that's a potentially bigger change.