lib/corres_method: more corresK map rules
Introduces F_all2, which allows us to have
a proper stateless precondition in corres_mapM_list_all2
(has access to all introduced goal parameters).

With it is the implicit assumption that the generated
condition can be phrased as a per-element property (although
it may talk about the list as a whole).