cspec: Reconfigure kernel build on sources change
Previously the kernel build would be reconfigured only if the explicit config file
passed in -C changed, but this file is very minimal and is only used when a variable
needs to be set to a non default value. As a result when default values are added
or changed a reconfigure will not be triggered.

This commit pays higher build times (by completely purging the build directory) in
order to more reliably capture reconfigures and rebuilds.