infoflow: add InfoFlow_Image_Toplevel
It's really tiring figuring out whether we loaded all of the right
InfoFlow theory files in jEdit. This file lists what "the theories for
InfoFlow" are and should be loaded instead.

ROOT file adjusted to target it instead of a bunch of files, some of
which already include some of the others.