rt ainvs: insert valid_ntfn_qs as precondition to send_signal_valid_sched
This predicate asserts that the threads in the queues of notification
objects either have no sched context, or the sched context is active,
budget_sufficient, and budget_ready. This is inserted as a precondition
to send_signal_valid_sched and the associated proofs are updated and
fixed.

Ideally this will become a part of valid_sched, this work is a first
step towards that.