rt ainvs: prove decode_inv_wf
This strengthens the precondition of `decode_inv_wf`, requiring that
extra caps are *equal* to the caps in the slots they came from, not just
`diminished` and `interrupt_derived`.

This strenghtening is fine, because `lookup_extra_caps` does not mask or
derive the caps it returns. It is also necessary to satisfy the
preconditions of lemmas about functions called by `decode_invocation`.
It might be possible to weaken the preconditions of those subordinate
lemmas, but given that `lookup_extra_caps` returns caps unchanged, there
is no reason to do this.

It would be nice to eradicate `diminished` and `interrupt_derived` from
the extra caps preconditions of all decode lemmas, but we don't attempt
that here.