c-parser: always emit long names.
Previously, the C parser would define locals differently depending on
the order they appear in the source (the first instance got a short
name, the second etc. got a longer one). This would sometimes make
things break when source was reordered.

Now, the C parser emits the long name for _every_ local, and emits an
abbreviation for backwards-compatibility and convenience for common
variables (like loop indexes `int i`).

Adjusts the Simpl syntax modifiers to work with abbreviations.

Modifies the VCG tactic to try and convert long-name bound variables in
the goal to their abbreviated names.