riscv aspec: change pte to store ppn instead of address
On other architectures, the address is a PTE stored using field_high and
thus retrieved as an aligned address. On RISCV we have a frame number
(referred to as PPN in some places) that is the address shifted down by
pt_bits.

This changes over the pte to use a ppn with a different number of bits,
and provides addr_from_ppn and addr_from_pte accessors, the latter being
an abbreviation.

Issues:
- "ppn" and "frame" show up in C, which should we use
- conversion functions take paddr, but are named with "addr": change
  naming to use paddr?
- we sanity check the number of bits in a ppn is word_bits - pt_bits,
  but in C that number subtracts another 8 bits, not clear why