sys-init: use executable variant of used_irq_list
This alternative formulation determines the set of used_irqs in
a way that is less roundabout and closer to the C implementation:
instead of  "every irq such that some object has an irqhandler cap
for it", it's "every irq such that there is an irq node for it".

For the proofs, we prove that the two formulations agree on well-formed
specs. Beyond using this fact, the old proofs need not change at all.

Incorporates some proof improvements suggested by @jalim and
shamelessly stolen.