rt ainvs: introduced valid_machine_time and cleaned up check_budget_valid_sched_fallout
valid_machine_time is an invariant stating that
last_machine_time (machine_state s) is bounded above by
-kernel_ticks-1 and is greater than or equal to cur_time s.
This invariant is required to prove the boundedness of certain
other values, adding this invariant removes an assumption in
rt_assumptions.

The first two assumptions in check_budget_valid_sched_fallout
were essentially contained in valid_refills and their existence
was a result of lazy coding in the past. I removed these assumptions
and incorporated valid_refills as preconditions in proofs where
required. This introduced two further sorries that require more work,
but are reasonable and expected to be true.